day2/RAG/requirements.txt
Flask==3.1.0
azure-identity==1.19.0
azure-search-documents==11.6.0b5
azure-core<=1.31.0
openai==1.57.2